show offer info and open website and maps

+++ Development/app/src/main/java/com/unisinos/ubiquitouscomputingtgb/OfferAdapter.java
@@ -55,8 +55,10 @@ class OfferAdapter extends BaseAdapter{
         textDealer.setText(offer.getDealer());
         textPrice.setText("U$ " + String.valueOf(offer.getPrice()));
         textOfferType.setText(offer.getOfferType());
-        textDistance.setText(String.valueOf(offer.getDistanceToUser()) + " meters");
+        textDistance.setText(String.valueOf((int)offer.getDistanceToUser()) + " meters");
 
         return convertView;
     }
+
+
 }
